requirements.txt
altgraph==0.17.4
anyio==4.10.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
attrs==25.3.0
babel==2.17.0
beautifulsoup4==4.13.5
bleach==6.2.0
certifi==2025.8.3
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.3
colorama==0.4.6
comm==0.2.3
debugpy==1.8.16
decorator==5.2.1
defusedxml==0.7.1
docx2pdf==0.1.8
et_xmlfile==2.0.0
executing==2.2.1
fastjsonschema==2.21.2
fqdn==1.5.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.7
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
lark==1.2.2
loguru==0.7.3
lxml==5.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.4
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.4.5
notebook_shim==0.2.4
numpy==2.3.2
openpyxl==3.1.5
packaging==24.2
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
pefile==2023.2.7
pillow==11.1.0
platformdirs==4.4.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psutil==7.0.0
pure_eval==0.2.3
pycparser==2.23
pycryptodome==3.23.0
Pygments==2.19.2
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
PyMuPDF==1.25.1
pypandoc==1.14
pyperclip==1.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
pyserial==3.5
PySide6==6.8.2
PySide6_Addons==6.8.2
PySide6_Essentials==6.8.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-json-logger==3.3.0
python-pptx==1.0.2
pytz==2025.2
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==3.0.0
PyYAML==6.0.2
pyzmq==27.1.0
referencing==0.36.2
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.27.1
Send2Trash==1.8.3
setuptools==75.7.0
shiboken6==6.8.2
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
TSMasterAPI==2024.8.22.1171
types-python-dateutil==2.9.0.20250822
typing_extensions==4.12.2
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.14
win32_setctime==1.2.0
XlsxWriter==3.2.0
